feat: delete post API
There is a one-to-many relationship in which, when the parent entity disappears, the child entities must all disappear, as in the cases of post and comment, comment and commet reply, and post and scrap post. In that case, I added Cascade.ALL to make the life cycles match. For example, when a post is deleted, comments, comment replies, and scrap posts are also deleted.