feat: add delete comment API
When a reply was made to a comment, instead of deleting the comment, the content of the post was changed to "삭제된 댓글입니다.". This means that if user delete a comment with a reply, the member(user) will still be able to access it. Later, i have to develop a logic that also disconnects the member.